requirements.txt
flask==2.3.3
opencv-python==4.8.1.78
face_recognition==1.3.0
numpy==1.26.4
Pillow==10.4.0
requests==2.32.3